add(lab09): Adding Lab09 deliverables

+++ labs/lab09_robots/src/robot.cc
@@ -32,8 +32,8 @@ void Robot::Update(double time) {
 	
 	double xvel, yvel, xnow, ynow, xprev, yprev;
 
-    xnow = origin_.x_ + 200.0 * cos(time * speed_);
-    ynow = origin_.y_ + 200.0 * sin(time * speed_);
+    xnow = (double) (origin_.x_ + 200.0 * cos(time * speed_));
+    ynow = (double) (origin_.y_ + 200.0 * sin(time * speed_));
     xprev = position_.x_;
     yprev = position_.y_;
